+++ src/utils/SorterUtils.js
@@ -18,11 +18,7 @@ export function sortObjects(objects, fields, filter, state, getMetaDataFunction,
                 continue;
             }
 
-            result = compareObjectsHierarchy(field, a, b, state, getMetaDataFunction, indented);
-
-            if (sortDirection === 'descending') {
-                result *= -1;
-            }
+            result = compareObjectsHierarchy(field, a, b, sortDirection, state, getMetaDataFunction, indented);
 
             if (result !== 0) {
                 break;
@@ -30,11 +26,11 @@ export function sortObjects(objects, fields, filter, state, getMetaDataFunction,
         }
 
         if (result === 0) {
-            result = compareObjectsHierarchy(fields.find(field => field.id === 'title'), a, b, state, getMetaDataFunction, indented);
+            result = compareObjectsHierarchy(fields.find(field => field.id === 'title'), a, b, 'ascending', state, getMetaDataFunction, indented);
         }
 
         if (result === 0) {
-            result = compareObjectsHierarchy(fields.find(field => field.id === 'id'), a, b, state, getMetaDataFunction, indented);
+            result = compareObjectsHierarchy(fields.find(field => field.id === 'id'), a, b, 'ascending', state, getMetaDataFunction, indented);
         }
 
         return result;